Move knoxite-lib into root folder and cmd into cmd/knoxite
