+++ chunkindex_test.go
