+++ cmd/knoxite/mount.go
@@ -3,7 +3,7 @@
 
 /*
  * knoxite
- *     Copyright (c) 2016-2017, Christian Muehlhaeuser <muesli@gmail.com>
+ *     Copyright (c) 2016-2020, Christian Muehlhaeuser <muesli@gmail.com>
  *
  *   For license see LICENSE
  */
@@ -23,7 +23,7 @@ import (
 	"github.com/spf13/cobra"
 	"golang.org/x/net/context"
 
-	knoxite "github.com/knoxite/knoxite/lib"
+	"github.com/knoxite/knoxite"
 )
 
 var (
