+++ decode.go
