+++ repository.go
@@ -1,267 +1,188 @@
 /*
  * knoxite
- *     Copyright (c) 2016-2017, Christian Muehlhaeuser <muesli@gmail.com>
+ *     Copyright (c) 2016-2018, Christian Muehlhaeuser <muesli@gmail.com>
  *
  *   For license see LICENSE
  */
 
-package main
+package knoxite
 
 import (
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
-	"strings"
-	"os"
-	"syscall"
+)
 
-	"github.com/klauspost/shutdown2"
-	"github.com/muesli/gotable"
-	"github.com/spf13/cobra"
-	"golang.org/x/crypto/ssh/terminal"
-	"github.com/muesli/crunchy"
+// A Repository is a collection of backup snapshots
+// MUST BE encrypted
+type Repository struct {
+	Version uint      `json:"version"`
+	Volumes []*Volume `json:"volumes"`
+	Paths   []string  `json:"storage"`
+	// Owner   string    `json:"owner"`
 
-	knoxite "github.com/knoxite/knoxite/lib"
+	backend  BackendManager
+	password string
+}
+
+// Const declarations
+const (
+	RepositoryVersion = 3
 )
 
 // Error declarations
 var (
-	ErrPasswordMismatch = errors.New("Passwords did not match")
-
-	repoCmd = &cobra.Command{
-		Use:   "repo",
-		Short: "manage repository",
-		Long:  `The repo command manages repositories`,
-		RunE:  nil,
-	}
-	repoInitCmd = &cobra.Command{
-		Use:   "init",
-		Short: "initialize a new repository",
-		Long:  `The init command initializes a new repository`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return executeRepoInit()
-		},
-	}
-	repoCatCmd = &cobra.Command{
-		Use:   "cat",
-		Short: "display repository information as JSON",
-		Long:  `The cat command displays the internal repository information as JSON`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return executeRepoCat()
-		},
-	}
-	repoInfoCmd = &cobra.Command{
-		Use:   "info",
-		Short: "display repository information",
-		Long:  `The info command displays the repository status & information`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return executeRepoInfo()
-		},
-	}
-	repoAddCmd = &cobra.Command{
-		Use:   "add <url>",
-		Short: "add another storage backend to a repository",
-		Long:  `The add command adds another storage backend to a repository`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 1 {
-				return fmt.Errorf("add needs a URL to be added")
-			}
-			return executeRepoAdd(args[0])
-		},
-	}
-	repoPackCmd = &cobra.Command{
-		Use:   "pack",
-		Short: "pack repository and release redundant data",
-		Long:  `The pack command deletes all unused data chunks from storage`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return executeRepoPack()
-		},
-	}
+	ErrRepositoryIncompatible = errors.New("The repository is not compatible with this version of Knoxite")
+	ErrOpenRepositoryFailed   = errors.New("Wrong password or corrupted repository")
+	ErrVolumeNotFound         = errors.New("Volume not found")
+	ErrSnapshotNotFound       = errors.New("Snapshot not found")
 )
 
-func init() {
-	repoCmd.AddCommand(repoInitCmd)
-	repoCmd.AddCommand(repoCatCmd)
-	repoCmd.AddCommand(repoInfoCmd)
-	repoCmd.AddCommand(repoAddCmd)
-	repoCmd.AddCommand(repoPackCmd)
-	RootCmd.AddCommand(repoCmd)
+// NewRepository returns a new repository
+func NewRepository(path, password string) (Repository, error) {
+	repository := Repository{
+		Version:  RepositoryVersion,
+		password: password,
+	}
+
+	backend, err := BackendFromURL(path)
+	if err != nil {
+		return repository, err
+	}
+	repository.backend.AddBackend(&backend)
+
+	err = repository.init()
+	return repository, err
 }
 
-func executeRepoInit() error {
-	// acquire a shutdown lock. we don't want these next calls to be interrupted
-	lock := shutdown.Lock()
-	if lock == nil {
-		return nil
+// OpenRepository opens an existing repository
+func OpenRepository(path, password string) (Repository, error) {
+	repository := Repository{
+		password: password,
 	}
-	defer lock()
 
-	r, err := newRepository(globalOpts.Repo, globalOpts.Password)
+	backend, err := BackendFromURL(path)
 	if err != nil {
-		return fmt.Errorf("Creating repository at %s failed: %v", globalOpts.Repo, err)
+		return repository, err
+	}
+	b, err := backend.LoadRepository()
+	if err != nil {
+		return repository, err
 	}
 
-	fmt.Printf("Created new repository at %s\n", (*r.BackendManager().Backends[0]).Location())
+	pipe, err := NewDecodingPipeline(CompressionNone, EncryptionAES, password)
+	if err != nil {
+		return repository, err
+	}
+	err = pipe.Decode(b, &repository)
+	if err != nil {
+		return repository, ErrOpenRepositoryFailed
+	}
+	if repository.Version < RepositoryVersion {
+		return repository, ErrRepositoryIncompatible
+	}
+
+	for _, url := range repository.Paths {
+		backend, berr := BackendFromURL(url)
+		if berr != nil {
+			return repository, berr
+		}
+		repository.backend.AddBackend(&backend)
+	}
+
+	return repository, err
+}
+
+// AddVolume adds a volume to a repository
+func (r *Repository) AddVolume(volume *Volume) error {
+	r.Volumes = append(r.Volumes, volume)
 	return nil
 }
 
-func executeRepoAdd(url string) error {
-	// acquire a shutdown lock. we don't want these next calls to be interrupted
-	lock := shutdown.Lock()
-	if lock == nil {
-		return nil
-	}
-	defer lock()
-
-	r, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
+// FindVolume finds a volume within a repository
+func (r *Repository) FindVolume(id string) (*Volume, error) {
+	if id == "latest" && len(r.Volumes) > 0 {
+		return r.Volumes[len(r.Volumes)-1], nil
 	}
 
-	backend, err := knoxite.BackendFromURL(url)
-	if err != nil {
-		return err
-	}
-	r.BackendManager().AddBackend(&backend)
-
-	err = r.Save()
-	if err != nil {
-		return err
-	}
-	fmt.Printf("Added %s to repository\n", backend.Location())
-	return nil
-}
-
-func executeRepoCat() error {
-	r, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
-	}
-
-	json, err := json.MarshalIndent(r, "", "    ")
-	if err != nil {
-		return err
-	}
-	fmt.Printf("%s\n", json)
-	return nil
-}
-
-func executeRepoPack() error {
-	r, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
-	}
-	index, err := knoxite.OpenChunkIndex(&r)
-	if err != nil {
-		return err
-	}
-
-	freedSize, err := index.Pack(&r)
-	if err != nil {
-		return err
-	}
-
-	err = index.Save(&r)
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("Freed storage space: %s\n", knoxite.SizeToString(freedSize))
-	return nil
-}
-
-func executeRepoInfo() error {
-	r, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
-	}
-
-	tab := gotable.NewTable([]string{"Storage URL", "Available Space"},
-		[]int64{-48, 15},
-		"No backends found.")
-
-	for _, be := range r.BackendManager().Backends {
-		space, _ := (*be).AvailableSpace()
-		tab.AppendRow([]interface{}{
-			(*be).Location(),
-			knoxite.SizeToString(space)})
-	}
-
-	_ = tab.Print()
-	return nil
-}
-
-func openRepository(path, password string) (knoxite.Repository, error) {
-	if password == "" {
-		var err error
-		password, err = readPassword("Enter password:")
-		if err != nil {
-			return knoxite.Repository{}, err
+	for _, volume := range r.Volumes {
+		if volume.ID == id {
+			return volume, nil
 		}
 	}
 
-	return knoxite.OpenRepository(path, password)
+	return &Volume{}, ErrVolumeNotFound
 }
 
-func newRepository(path, password string) (knoxite.Repository, error) {
-	if password == "" {
-		var err error
-		password, err = readPasswordTwice("Enter a password to encrypt this repository with:", "Confirm password:")
-		if err != nil {
-			return knoxite.Repository{}, err
+// FindSnapshot finds a snapshot within a repository
+func (r *Repository) FindSnapshot(id string) (*Volume, *Snapshot, error) {
+	if id == "latest" {
+		latestVolume := &Volume{}
+		latestSnapshot := &Snapshot{}
+		found := false
+		for _, volume := range r.Volumes {
+			for _, snapshotID := range volume.Snapshots {
+				snapshot, err := volume.LoadSnapshot(snapshotID, r)
+				if err == nil {
+					if !found || snapshot.Date.Sub(latestSnapshot.Date) > 0 {
+						latestSnapshot = snapshot
+						latestVolume = volume
+						found = true
+					}
+				}
+			}
 		}
-	}
 
-	return knoxite.NewRepository(path, password)
-}
-
-func readPassword(prompt string) (string, error) {
-	var tty io.WriteCloser
-	tty, err := os.OpenFile("/dev/tty", os.O_WRONLY, 0)
-	if err != nil {
-		tty = os.Stdout
+		if found {
+			return latestVolume, latestSnapshot, nil
+		}
 	} else {
-		defer tty.Close()
+		for _, volume := range r.Volumes {
+			snapshot, err := volume.LoadSnapshot(id, r)
+			if err == nil {
+				return volume, snapshot, err
+			}
+		}
 	}
 
-	fmt.Fprint(tty, prompt+" ")
-	buf, err := terminal.ReadPassword(int(syscall.Stdin))
-	fmt.Fprintln(tty)
-
-	return string(buf), err
+	return &Volume{}, &Snapshot{}, ErrSnapshotNotFound
 }
 
-func readPasswordTwice(prompt, promptConfirm string) (string, error) {
-	pw, err := readPassword(prompt)
-	if err != nil {
-		return pw, err
-	}
-
-	crunchErr := crunchy.NewValidator().Check(pw)
-	if crunchErr != nil {
-		fmt.Printf("Password is considered unsafe: %v\n", crunchErr)
-		fmt.Printf("Are you sure you want to use this password (y/N)?: ")
-		var buf string
-		_, err = fmt.Scan(&buf)
-		if err != nil {
-			return pw, err
-		}
-
-		buf = strings.TrimSpace(buf)
-		buf = strings.ToLower(buf)
-		if buf != "y" {
-			return pw, crunchErr
+// IsEmpty returns true if there a no snapshots stored in a repository
+func (r *Repository) IsEmpty() bool {
+	for _, volume := range r.Volumes {
+		if len(volume.Snapshots) > 0 {
+			return false
 		}
 	}
 
-	pwconfirm, err := readPassword(promptConfirm)
+	return true
+}
+
+// BackendManager returns the repository's BackendManager
+func (r *Repository) BackendManager() *BackendManager {
+	return &r.backend
+}
+
+// Init creates a new repository
+func (r *Repository) init() error {
+	err := r.backend.InitRepository()
 	if err != nil {
-		return pw, err
-	}
-	if pw != pwconfirm {
-		return pw, ErrPasswordMismatch
+		return err
 	}
 
-	return pw, nil
+	return r.Save()
+}
+
+// Save writes a repository's metadata
+func (r *Repository) Save() error {
+	r.Paths = r.backend.Locations()
+
+	pipe, err := NewEncodingPipeline(CompressionNone, EncryptionAES, r.password)
+	if err != nil {
+		return err
+	}
+	b, err := pipe.Encode(r)
+	if err != nil {
+		return err
+	}
+	return r.backend.SaveRepository(b)
 }
