+++ scanner.go
