+++ snapshot.go
@@ -1,126 +1,227 @@
 /*
  * knoxite
- *     Copyright (c) 2016-2017, Christian Muehlhaeuser <muesli@gmail.com>
+ *     Copyright (c) 2016-2018, Christian Muehlhaeuser <muesli@gmail.com>
  *
  *   For license see LICENSE
  */
 
-package main
+package knoxite
 
 import (
-	"fmt"
+	"math"
+	"os"
+	"path/filepath"
+	"strings"
+	"sync"
+	"time"
 
-	"github.com/muesli/gotable"
-	"github.com/spf13/cobra"
-
-	knoxite "github.com/knoxite/knoxite/lib"
+	uuid "github.com/nu7hatch/gouuid"
 )
 
-var (
-	snapshotCmd = &cobra.Command{
-		Use:   "snapshot",
-		Short: "manage snapshots",
-		Long:  `The snapshot command manages snapshots`,
-		RunE:  nil,
-	}
-	snapshotListCmd = &cobra.Command{
-		Use:   "list <volume>",
-		Short: "list all snapshots inside a volume",
-		Long:  `The list command lists all snapshots stored in a volume`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 1 {
-				return fmt.Errorf("list needs a volume ID to work on")
-			}
-			return executeSnapshotList(args[0])
-		},
-	}
-	snapshotRemoveCmd = &cobra.Command{
-		Use:   "remove <snapshot>",
-		Short: "remove a snapshot",
-		Long:  `The remove command deletes a snapshot from a volume`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 1 {
-				return fmt.Errorf("remove needs a snapshot ID to work on")
-			}
-			return executeSnapshotRemove(args[0])
-		},
-	}
-)
+// A Snapshot is a compilation of one or many archives
+// MUST BE encrypted
+type Snapshot struct {
+	mut sync.Mutex
 
-func init() {
-	snapshotCmd.AddCommand(snapshotListCmd)
-	snapshotCmd.AddCommand(snapshotRemoveCmd)
-	RootCmd.AddCommand(snapshotCmd)
+	ID          string              `json:"id"`
+	Date        time.Time           `json:"date"`
+	Description string              `json:"description"`
+	Stats       Stats               `json:"stats"`
+	Archives    map[string]*Archive `json:"items"`
 }
 
-func executeSnapshotRemove(snapshotID string) error {
-	repository, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
-	}
-	chunkIndex, err := knoxite.OpenChunkIndex(&repository)
-	if err != nil {
-		return err
+// NewSnapshot creates a new snapshot
+func NewSnapshot(description string) (*Snapshot, error) {
+	snapshot := Snapshot{
+		Date:        time.Now(),
+		Description: description,
+		Archives:    make(map[string]*Archive),
 	}
 
-	volume, snapshot, err := repository.FindSnapshot(snapshotID)
+	u, err := uuid.NewV4()
 	if err != nil {
-		return err
+		return &snapshot, err
 	}
+	snapshot.ID = u.String()[:8]
 
-	err = volume.RemoveSnapshot(snapshot.ID)
-	if err != nil {
-		return err
-	}
-
-	chunkIndex.RemoveSnapshot(snapshot.ID)
-	err = chunkIndex.Save(&repository)
-	if err != nil {
-		return err
-	}
-
-	err = repository.Save()
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("Snapshot %s removed: %s\n", snapshot.ID, snapshot.Stats.String())
-	fmt.Println("Do not forget to run 'repo pack' to delete un-referenced chunks and free up storage space!")
-	return nil
+	return &snapshot, nil
 }
 
-func executeSnapshotList(volID string) error {
-	repository, err := openRepository(globalOpts.Repo, globalOpts.Password)
-	if err != nil {
-		return err
-	}
+func (snapshot *Snapshot) gatherTargetInformation(cwd string, paths []string, excludes []string, out chan ArchiveResult) {
+	var wg sync.WaitGroup
+	for _, path := range paths {
+		c := findFiles(path, excludes)
 
-	volume, err := repository.FindVolume(volID)
-	if err != nil {
-		return err
-	}
+		for result := range c {
+			if result.Error == nil {
+				rel, err := filepath.Rel(cwd, result.Archive.Path)
+				if err == nil && !strings.HasPrefix(rel, "../") {
+					result.Archive.Path = rel
+				}
+				if isSpecialPath(result.Archive.Path) {
+					continue
+				}
 
-	tab := gotable.NewTable([]string{"ID", "Date", "Original Size", "Storage Size", "Description"},
-		[]int64{-8, -19, 13, 12, -48}, "No snapshots found. This volume is empty.")
-	totalSize := uint64(0)
-	totalStorageSize := uint64(0)
+				snapshot.mut.Lock()
+				snapshot.Stats.Size += result.Archive.Size
+				switch result.Archive.Type {
+				case Directory:
+					snapshot.Stats.Dirs++
+				case File:
+					snapshot.Stats.Files++
+				case SymLink:
+					snapshot.Stats.SymLinks++
+				}
+				snapshot.mut.Unlock()
+			}
 
-	for _, snapshotID := range volume.Snapshots {
-		snapshot, err := volume.LoadSnapshot(snapshotID, &repository)
-		if err != nil {
-			return err
+			wg.Add(1)
+			go func(r ArchiveResult) {
+				out <- r
+				wg.Done()
+			}(result)
 		}
-		tab.AppendRow([]interface{}{
-			snapshot.ID,
-			snapshot.Date.Format(timeFormat),
-			knoxite.SizeToString(snapshot.Stats.Size),
-			knoxite.SizeToString(snapshot.Stats.StorageSize),
-			snapshot.Description})
-		totalSize += snapshot.Stats.Size
-		totalStorageSize += snapshot.Stats.StorageSize
 	}
 
-	tab.SetSummary([]interface{}{"", "", knoxite.SizeToString(totalSize), knoxite.SizeToString(totalStorageSize), ""})
-	_ = tab.Print()
-	return nil
+	wg.Wait()
+	close(out)
+}
+
+// Add adds a path to a Snapshot
+func (snapshot *Snapshot) Add(cwd string, paths []string, excludes []string, repository Repository, chunkIndex *ChunkIndex, compress, encrypt uint16, dataParts, parityParts uint) chan Progress {
+	progress := make(chan Progress)
+	fwd := make(chan ArchiveResult)
+
+	go snapshot.gatherTargetInformation(cwd, paths, excludes, fwd)
+
+	go func() {
+		for result := range fwd {
+			if result.Error != nil {
+				p := newProgressError(result.Error)
+				progress <- p
+				break
+			}
+
+			archive := result.Archive
+			rel, err := filepath.Rel(cwd, archive.Path)
+			if err == nil && !strings.HasPrefix(rel, "../") {
+				archive.Path = rel
+			}
+			if isSpecialPath(archive.Path) {
+				continue
+			}
+
+			p := newProgress(archive)
+			snapshot.mut.Lock()
+			p.TotalStatistics = snapshot.Stats
+			snapshot.mut.Unlock()
+			progress <- p
+
+			if archive.Type == File {
+				dataParts = uint(math.Max(1, float64(dataParts)))
+				chunkchan, err := chunkFile(archive.Path, compress, encrypt, repository.password, int(dataParts), int(parityParts))
+				if err != nil {
+					if os.IsNotExist(err) {
+						// if this file has already been deleted before we could backup it, we can gracefully ignore it and continue
+						continue
+					}
+					p = newProgressError(err)
+					progress <- p
+					break
+				}
+				archive.Encrypted = encrypt
+				archive.Compressed = compress
+
+				for cd := range chunkchan {
+					if cd.Error != nil {
+						p = newProgressError(err)
+						progress <- p
+						close(progress)
+						return
+					}
+					chunk := cd.Chunk
+					// fmt.Printf("\tSplit %s (#%d, %d bytes), compression: %s, encryption: %s, hash: %s\n", id.Path, cd.Num, cd.Size, CompressionText(cd.Compressed), EncryptionText(cd.Encrypted), cd.Hash)
+
+					// store this chunk
+					n, err := repository.backend.StoreChunk(chunk)
+					if err != nil {
+						p = newProgressError(err)
+						progress <- p
+						close(progress)
+						return
+					}
+
+					// release the memory, we don't need the data anymore
+					chunk.Data = &[][]byte{}
+
+					archive.Chunks = append(archive.Chunks, chunk)
+					archive.StorageSize += n
+
+					p.CurrentItemStats.StorageSize = archive.StorageSize
+					p.CurrentItemStats.Transferred += uint64(chunk.OriginalSize)
+					snapshot.Stats.Transferred += uint64(chunk.OriginalSize)
+					snapshot.Stats.StorageSize += n
+
+					snapshot.mut.Lock()
+					p.TotalStatistics = snapshot.Stats
+					snapshot.mut.Unlock()
+					progress <- p
+				}
+			}
+
+			snapshot.AddArchive(archive)
+			chunkIndex.AddArchive(archive, snapshot.ID)
+		}
+		close(progress)
+	}()
+
+	return progress
+}
+
+// Clone clones a snapshot
+func (snapshot *Snapshot) Clone() (*Snapshot, error) {
+	s, err := NewSnapshot(snapshot.Description)
+	if err != nil {
+		return s, err
+	}
+
+	s.Stats = snapshot.Stats
+	s.Archives = snapshot.Archives
+
+	return s, nil
+}
+
+// openSnapshot opens an existing snapshot
+func openSnapshot(id string, repository *Repository) (*Snapshot, error) {
+	snapshot := Snapshot{
+		Archives: make(map[string]*Archive),
+	}
+	b, err := repository.backend.LoadSnapshot(id)
+	if err != nil {
+		return &snapshot, err
+	}
+	pipe, err := NewDecodingPipeline(CompressionLZMA, EncryptionAES, repository.password)
+	if err != nil {
+		return &snapshot, err
+	}
+	err = pipe.Decode(b, &snapshot)
+	return &snapshot, err
+}
+
+// Save writes a snapshot's metadata
+func (snapshot *Snapshot) Save(repository *Repository) error {
+	pipe, err := NewEncodingPipeline(CompressionLZMA, EncryptionAES, repository.password)
+	if err != nil {
+		return err
+	}
+	b, err := pipe.Encode(snapshot)
+	if err != nil {
+		return err
+	}
+	return repository.backend.SaveSnapshot(snapshot.ID, b)
+}
+
+// AddArchive adds an archive to a snapshot
+func (snapshot *Snapshot) AddArchive(archive *Archive) {
+	snapshot.Archives[archive.Path] = archive
 }
