+++ stat_generic.go
