+++ stat_windows.go
