+++ statistics_test.go
