+++ storage_local_windows.go
